requirements.txt
google-genai==1.47.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests~=2.32.0
uvicorn==0.38.0
pydantic~=2.12.3
fastapi~=0.120.2
google~=3.0.0
httpx~=0.28.1

